+++ CHANGELOG.md
@@ -27,3 +27,5 @@
 - updated logo and toggle icons
 - added option to switch between WebGL and 2D canvas
 - updated asteroids and enemies (images, health, strength, fire rate, ...)
+- added background music and option to toggle it on/off
+- increased brightness of toggle icons